pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "email-triage"
version = "0.1.0"
description = "Email Triage OpenEnv Environment"
dependencies = [
    "pydantic",
    "openenv-core",
    "fastapi",
    "uvicorn",
    "openai"
]

[project.scripts]
server = "server.app:app"

